BROKEN WIP: working on Secp256k1 uncompressed
