+++ apple/Tests/SargonTests/TestCases/Crypto/Secp256k1PublicKeyTests.swift
@@ -1,3 +1,4 @@
+@testable import Sargon
 
 final class Secp256k1PublicKeyTests: PublicKeyTest<Secp256k1PublicKey> {
 	func test_from_compressed() throws {
@@ -9,4 +10,18 @@ final class Secp256k1PublicKeyTests: PublicKeyTest<Secp256k1PublicKey> {
 		// from K1: https://github.com/Sajjon/K1/blob/main/Tests/K1Tests/TestCases/Keys/PublicKey/PublicKeyImportTests.swift#L48
 		XCTAssertNoThrow(try SUT(hex: "040202020202020202020202020202020202020202020202020202020202020202415456f0fc01d66476251cab4525d9db70bfec652b2d8130608675674cde64b2"))
 	}
+    
+    func test_uncompressed_and_compresses_equals() throws {
+        try XCTAssertEqual(
+            SUT(hex: "040202020202020202020202020202020202020202020202020202020202020202415456f0fc01d66476251cab4525d9db70bfec652b2d8130608675674cde64b2"),
+            SUT(hex: "020202020202020202020202020202020202020202020202020202020202020202")
+        )
+    }
+    
+    func test_uncompressed_from_compressed() throws {
+        try XCTAssertEqual(
+            SUT(hex: "020202020202020202020202020202020202020202020202020202020202020202").uncompressedData.hex,
+            "040202020202020202020202020202020202020202020202020202020202020202415456f0fc01d66476251cab4525d9db70bfec652b2d8130608675674cde64b2"
+        )
+    }
 }
