+++ src/core/types/keys/secp256k1/mod.rs
@@ -1,5 +1,7 @@
 mod private_key;
 mod public_key;
+mod public_key_uniffi_fn;
 
 pub use private_key::*;
 pub use public_key::*;
+pub use public_key_uniffi_fn::*;
